add watchlist recommendations

+++ back-end/db/migrations/0001_create_movie_tables.up.sql
@@ -2,6 +2,6 @@ CREATE TABLE movie(
   id serial PRIMARY KEY,
   jellyfin_id VARCHAR(255) NOT NULL UNIQUE,
   title VARCHAR(255) NOT NULL,
-  production_year SMALLINT NOT NULL,
+  production_year INT NOT NULL,
   community_rating DECIMAL not null
 )
